DSS-576: Save on board quantity
save on board quantity, get Saved OBQ data

type: feature

related issues: DSS-574

+++ api/cp-dss-api/loadable-study/src/main/java/com/cpdss/loadablestudy/entity/OnBoardQuantity.java
@@ -28,10 +28,10 @@ public class OnBoardQuantity extends EntityDoc {
   private Long portId;
 
   @Column(name = "tank_xid")
-  private Integer tankId;
+  private Long tankId;
 
   @Column(name = "cargo_xid")
-  private Integer cargoId;
+  private Long cargoId;
 
   @Column(name = "sounding")
   private BigDecimal sounding;
